Refactoring

+++ Calculator/CalculatorTests.cs
@@ -6,10 +6,10 @@ namespace Calculator
     [TestFixture]
     public class CalculatorTests
     {
-        [TestCase(0.0, 1.0, 1.0)]
-        [TestCase(12.0,3.0,123.0)]
-        [TestCase(0.0,0.0,0.0)]
-        public void Anhängen(double operand, double ziffer, double expected)
+        [TestCase(0.0, 1, 1.0)]
+        [TestCase(12.0,3,123.0)]
+        [TestCase(0.0,0,0.0)]
+        public void Anhängen(double operand, int ziffer, double expected)
         {
             var calculator = new Calculator(new Calculator.State {Operand = operand});
             calculator.Anhängen(ziffer);
@@ -20,8 +20,8 @@ namespace Calculator
         public void AnhängenSequenziell()
         {
             var calculator = new Calculator(new Calculator.State { Operand = 0});
-            calculator.Anhängen(0.0);
-            calculator.Anhängen(0.0);
+            calculator.Anhängen(0);
+            calculator.Anhängen(0);
             Assert.AreEqual(0.0, calculator.Operand);
         }
